+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## Unreleased
 
+## [0.99.0](https://github.com/returntocorp/semgrep/releases/tag/v0.99.0) - 2022-06-22
+
 ## Added
 
 - taint-mode: New experimental `pattern-propagators` feature that allows to specify